pyproject.toml
[tool.poetry]
name = "zmq-ch4"
version = "0.1.0"
description = "ZeroMQ Book ch4 Examples"
authors = ["Jay R Bolton <jayrbolton@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pyzmq = "^19.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
